backend/pyproject.toml
[project]
name = "viet-advisory-orchestrator"
version = "0.1.0"
description = "AI Co-Pilot for Relationship Managers - Swin Hackathon 2026"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-aws>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.35.0",
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
